Excess of water

The function excessWater() processes the excess of water. It follows the expression of Skaugen and Onof (2013) and Skaugen,Peerebom and Nilsson(2015) if (G(t)+Z(t))/D(t) >R : X(t) = G(t) + Z(t) -R*D(t) if (G(t)+Z(t))/D(t) <R : X(t) = 0

excessWater(G, Z, D, R)

Arguments

G

Volume of received moisture input of rain and snowmelt a time t

Z

Actual water volume present in the soil moisture zone

D

the potential volumee of water that is needed for complete saturation at time t

R

Ratio defining field capacity (fracion of D)

Examples

## Not run: ------------------------------------
# excessWater()
## ---------------------------------------------